requirements.txt
adal==1.2.7
aiohttp==3.7.4.post0
aiounittest==1.3.0
applicationinsights==0.11.9
arrow==1.2.3
asgiref==3.5.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==22.1.0
azure-cognitiveservices-language-luis==0.2.0
azure-common==1.1.28
azure-core==1.26.1
azure-cosmos==3.2.0
azure-storage-blob==12.7.0
azure-storage-queue==12.1.5
Babel==2.9.1
binaryornot==0.4.4
botbuilder-ai==4.14.3
botbuilder-applicationinsights==4.14.3
botbuilder-azure==4.14.3
botbuilder-core==4.14.3
botbuilder-dialogs==4.14.3
botbuilder-integration-aiohttp==4.14.3
botbuilder-schema==4.14.3
botframework-connector==4.14.3
botframework-streaming==4.14.3
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cookiecutter==1.7.0
cryptography==38.0.3
datedelta==1.4
Django==3.2.16
djangorestframework==3.10.3
emoji==1.7.0
Flask==1.1.1
future==0.18.2
grapheme==0.6.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-time==0.2.0
jsonpickle==1.4.2
MarkupSafe==2.1.1
msal==1.20.0
msrest==0.6.21
msrestazure==0.6.4
multidict==6.0.2
multipledispatch==0.6.0
oauthlib==3.2.2
poyo==0.5.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
regex==2019.8.19
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
typing-extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
whichcraft==0.6.1
wrapt==1.14.1
yarl==1.4.2

